requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
ph-py==0.1
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
urllib3==1.22
